+++ src/actor/kernel/kernel.rs
@@ -394,7 +394,7 @@ impl<Extra: ExtraData> Kernel<Extra> {
 	/// Handler to when `Audio` messages us.
 	fn wrote_audio_buffer(&mut self, time: WroteAudioBuffer) {
 		// Calculate total time elapsed.
-		let elapsed = time.seconds as f64 + time.frac;
+		let elapsed = time.seconds as f32 + time.frac as f32;
 
 		// Update the `AudioState`.
 		self.w.add_commit_push(|w, _| {
